% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartik.R
\name{bartik}
\alias{bartik}
\title{Bartik Instrumental Regression}
\usage{
bartik()
}
\value{
\code{bartik} returns the results of the estimation.
}
\description{
\code{bartik} is used to fit a Bartik Instrumental-Variable regression.
}
